refactor ui structure
