+++ ui/src/features/AuthenticationWrapper/index.tsx
@@ -3,7 +3,7 @@ import { ReactNode } from "react";
 
 import { useAuthenticationRequired } from "@flow/lib/auth";
 
-const AuthenticationRequiredPage: React.FC<{ children?: ReactNode }> = ({ children }) => {
+const AuthenticationWrapper: React.FC<{ children?: ReactNode }> = ({ children }) => {
   const [isAuthenticated] = useAuthenticationRequired(); // TODO: show error
   return isAuthenticated && children ? <>{children}</> : null;
 };
@@ -12,4 +12,4 @@ const withAuthorisation = (): ((props: any) => React.FC<any>) => {
   return withAuthenticationRequired as unknown as (props: any) => React.FC<any>;
 };
 
-export default withAuthorisation()(AuthenticationRequiredPage);
+export default withAuthorisation()(AuthenticationWrapper);
