+++ ui/src/features/Editor/components/BottomPanel/components/DataTable/index.tsx
