+++ ui/src/features/Editor/components/BottomPanel/components/Map/index.tsx
