+++ ui/src/features/Editor/components/Canvas/components/Nodes/NoteNode/index.tsx
