+++ ui/src/features/Editor/components/Canvas/components/Nodes/WriterNode.tsx
