+++ ui/src/features/Editor/components/Canvas/components/index.ts
@@ -1,0 +1,3 @@
+export * from "./Nodes";
+export { default as CustomConnectionLine, connectionLineStyle } from "./CustomConnectionLine";
+export { default as CustomEdge, edgeTypes } from "./CustomEdge";
