--- ui/src/features/Editor/components/Nodes/GeneralNode/types.ts
@@ -1,3 +1,0 @@
-export type NodeType = "reader" | "transformer" | "writer" | "note" | "batch";
-
-export type NodePosition = { x: number; y: number };
