+++ ui/src/features/Editor/components/index.ts
@@ -1,6 +1,6 @@
-export * from "./CustomConnectionLine";
-export * from "./Nodes";
 export * from "./OverlayUI";
 export { default as Canvas } from "./Canvas";
 export * from "./WorkflowTabs";
-export { default as ParamEditor } from "./ParamEditor";
+export * from "./BottomPanel";
+export * from "./LeftPanel";
+export * from "./RightPanel";
