+++ ui/src/features/Runs/index.tsx
@@ -3,12 +3,11 @@ import { useNavigate, useParams } from "@tanstack/react-router";
 import { useState } from "react";
 
 import { Button } from "@flow/components";
+import { TopNavigation } from "@flow/features/TopNavigation";
 import { useT } from "@flow/lib/i18n";
 import { runs as mockRuns } from "@flow/mock_data/runsData";
 import { useCurrentWorkspace } from "@flow/stores";
 import { Run } from "@flow/types";
-
-import { TopNavigation } from "../TopNavigation";
 
 import { ManualRun } from "./components";
 import { StatusContent } from "./components/StatusContent";
