--- ui/src/pages/Dashboard.tsx
@@ -1,5 +1,0 @@
-import { Dashboard } from "@flow/features/Dashboard";
-
-const DashboardPage = () => <Dashboard />;
-
-export { DashboardPage };
