--- ui/src/pages/WorkspaceRuns.tsx
@@ -1,7 +1,0 @@
-import { Runs } from "@flow/features/Runs";
-
-function WorkspaceRuns() {
-  return <Runs />;
-}
-
-export { WorkspaceRuns };
