--- ui/src/pages/WorkspaceSettings.tsx
@@ -1,7 +1,0 @@
-import { WorkspaceSettings as Settings } from "@flow/features/WorkspaceSettings";
-
-function WorkspaceSettings() {
-  return <Settings />;
-}
-
-export { WorkspaceSettings };
