--- ui/src/pages/index.ts
@@ -1,5 +1,0 @@
-export * from "./IndexLoadingPage";
-export * from "./Editor";
-export * from "./WorkspaceRuns";
-export * from "./WorkspaceSettings";
-export * from "./Dashboard";
