+++ ui/src/routes/workspace_.$workspaceId_.project.$projectId.lazy.tsx
@@ -1,7 +1,9 @@
 import { createLazyFileRoute } from "@tanstack/react-router";
+import { ReactFlowProvider } from "@xyflow/react";
 
+import Canvas from "@flow/features/Editor";
 import { ProjectIdWrapper, WorkspaceIdWrapper } from "@flow/features/PageWrapper";
-import { Editor } from "@flow/pages";
+import { useCurrentProject } from "@flow/stores";
 
 export const Route = createLazyFileRoute("/workspace/$workspaceId/project/$projectId")({
   component: () => (
@@ -12,3 +14,15 @@ export const Route = createLazyFileRoute("/workspace/$workspaceId/project/$proje
     </WorkspaceIdWrapper>
   ),
 });
+
+function Editor() {
+  const [currentProject] = useCurrentProject();
+
+  return (
+    <div className="flex flex-col bg-zinc-900 text-zinc-300 h-screen">
+      <ReactFlowProvider>
+        <Canvas workflows={currentProject?.workflows} />
+      </ReactFlowProvider>
+    </div>
+  );
+}
