+++ ui/src/routes/workspace_.$workspaceId_.settings.$tab.lazy.tsx
@@ -1,12 +1,12 @@
 import { createLazyFileRoute } from "@tanstack/react-router";
 
 import { WorkspaceIdWrapper } from "@flow/features/PageWrapper";
-import { WorkspaceSettings } from "@flow/pages";
+import { WorkspaceSettings as Settings } from "@flow/features/WorkspaceSettings";
 
 export const Route = createLazyFileRoute("/workspace/$workspaceId/settings/$tab")({
   component: () => (
     <WorkspaceIdWrapper>
-      <WorkspaceSettings />
+      <Settings />
     </WorkspaceIdWrapper>
   ),
 });
